Extract pin dialog
